+++ docs/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - Added support for Django 4.2 (#1570)
+- Refactoring and fix to support filtering exports (#1579)
 
 
 3.2.0 (2023-04-12)