+++ docs/changelog.rst
@@ -6,8 +6,7 @@ Changelog
 
 - Added support for Django 4.2 (#1570)
 - Refactoring and fix to support filtering exports (#1579)
-- pass `row` dict to :meth:`~import_export.resources.Resource.after_import_instance` (#1584)
-
+- pass `original` object to :meth:`~import_export.resources.Resource.after_import_row` (#1584)
 
 3.2.0 (2023-04-12)
 ------------------